# docs/swagger.yaml
definitions:
...
  domain.EventPay:
    properties:
      event_id:
        type: integer
      id:
        type: integer
      image_proof_payment:
        type: string
      net_amount:
        type: number
      order_no:
        type: string
      registration_event:
        $ref: '#/definitions/domain.RegistrationEvent'
      registration_event_id:
        type: integer
      status:
        type: string
    type: object
...
  domain.HttpResponse:
    properties:
      code:
        type: integer
      data: {}
      limit:
        type: integer
      message:
        type: string
      page:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  domain.Login:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.PayProcessPayload:
    properties:
      note:
        type: string
      order_no:
        type: string
      status:
        type: string
    type: object
...
  domain.RegisterStatusResponse:
    properties:
      order_no:
        type: string
      string:
        type: string
    type: object
  domain.RegistrationEvent:
    properties:
      created_at:
        type: string
      created_by_user_id:
        type: integer
      deleted_at:
        type: string
      deleted_by_user_id:
        type: integer
      email:
        type: string
      event_id:
        description: lock event
        type: integer
      id:
        type: integer
      image_proof_payment:
        type: string
      name:
        type: string
      order_no:
        type: string
      payment_date:
        type: string
      phone_number:
        type: string
      status:
        description: register, pay, approve/cancel/decline
        type: string
      up_to_you:
        type: string
      updated_at:
        type: string
      updated_by_user_id:
        type: integer
      user_id:
        description: lock user
        type: string
    type: object
...
  domain.User:
    properties:
      address:
        type: string
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      fullname:
        type: string
      gender:
        type: string
      github:
        type: string
      id:
        type: integer
      linkedin:
        type: string
      password:
        type: string
      personal_web:
        type: string
      phone_number:
        type: string
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  domain.UserUpdateProfile:
    properties:
      address:
        type: string
      date_of_birth:
        type: string
      fullname:
        type: string
      gender:
        type: string
      github:
        type: string
      linkedin:
        type: string
      personal_web:
        type: string
      phone_number:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: hammercode.org
    name: API Support
    url: http://hammercode.org/support
  description: This is the API documentation for this service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Golang API (LMS-BE)
  version: "1.0"
paths:
...
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: This endpoint use to logout
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Logout
      tags:
      - Auth
...
  /api/v1/events/pays:
    get:
      consumes:
      - application/json
      description: This endpoint use to get list of event pay
      parameters:
      - description: string
        in: query
        name: limit
        required: true
        type: string
      - description: string
        in: query
        name: page
        required: true
        type: string
      - description: string
        in: query
        name: event_id
        type: string
      - description: string
        in: query
        name: start_date
        type: string
      - description: string
        in: query
        name: end_date
        type: string
      - description: string
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.EventPay'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Get List event pay
      tags:
      - Event
    post:
      consumes:
      - application/json
      description: This endpoint use to confirm payment
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.PayProcessPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Pay Process
      tags:
      - Event
  /api/v1/events/registartions/:order_no:
    get:
      consumes:
      - application/json
      description: This endpoint use to check registration status
      parameters:
      - description: ABCXX
        in: path
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.RegisterStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Register Status
      tags:
      - Event
...
  /api/v1/users:
    delete:
      consumes:
      - application/json
      description: This endpoint use to delete user
      parameters:
      - description: string
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Delete User
      tags:
      - User
    get:
      consumes:
      - application/json
      description: This endpoint use to get current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Get User Profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: This endpoint use to update profile user
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UserUpdateProfile'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.HttpResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.HttpResponse'
      summary: Update User
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"